+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.248.1 - 2025-05-28
+
+- chore: sync with main repo (#1982)
+
+## 1.248.0 - 2025-05-27
+
+- feat: Add "defaults" to config (#1978)
+
+## 1.247.0 - 2025-05-27
+
+- fix: getSurveySeen and waitPeriod checks (#1979)
+
 ## 1.246.0 - 2025-05-22
 
 - fix: Block newer google bots (#1974)